refactor: pricing page cleanup
