+++ components/home/pricing/PricingDiscounts.tsx
@@ -1,0 +1,66 @@
+import { Button } from "@/components/ui/button";
+import Link from "next/link";
+
+const discounts = [
+  {
+    name: "Early-stage startups",
+    description: "50% off, first year",
+    cta: {
+      text: "Learn more and apply",
+      href: "/startups",
+    },
+  },
+  {
+    name: "Research / Students",
+    description: "Up to 100% off, limits apply",
+    cta: {
+      text: "Learn more and apply",
+      href: "/research",
+    },
+  },
+  {
+    name: "Open-source projects",
+    description: "USD 300 in credits / month, first year",
+  },
+];
+
+export const PricingDiscounts = () => (
+  <div className="mx-auto max-w-7xl px-6 lg:px-8 pt-8 mt-8" id="discounts">
+    <div className="mx-auto max-w-4xl">
+      <h2 className="text-2xl font-bold leading-10 tracking-tight text-primary">
+        Discounts
+      </h2>
+      <div className="mt-3 grid grid-cols-1 gap-4 sm:grid-cols-3">
+        {discounts.map((discount) => (
+          <div
+            key={discount.name}
+            className="rounded-lg border border-gray-200 p-4 shadow-sm"
+          >
+            <dt className="text-base font-semibold leading-7 text-primary">
+              {discount.name}
+            </dt>
+            <dd className="mt-2 text-sm leading-7 text-primary/60">
+              {discount.description}
+            </dd>
+            {discount.cta && (
+              <Button size="sm" variant="secondary" asChild className="mt-2">
+                <Link href={discount.cta.href} target="_blank">
+                  {discount.cta.text}
+                </Link>
+              </Button>
+            )}
+          </div>
+        ))}
+      </div>
+      <p className="mt-6 leading-8 text-primary/60">
+        Reach out to{" "}
+        <Link href="mailto:support@langfuse.com" className="underline">
+          support@langfuse.com
+        </Link>{" "}
+        to apply for a discount. We want all startups, educational users,
+        non-profits and open source projects to build with Langfuse and are
+        happy to work with you to make that happen.
+      </p>
+    </div>
+  </div>
+);
