+++ components/home/pricing/PricingTable.tsx
@@ -7,20 +7,15 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Check, Plus, Minus, ExternalLink, InfoIcon } from "lucide-react";
-import { Disclosure } from "@headlessui/react";
+import { Check, ExternalLink, InfoIcon } from "lucide-react";
 import Link from "next/link";
-import { Header } from "../Header";
-import { HomeSection } from "./components/HomeSection";
 import { cn } from "@/lib/utils";
-import { Tabs, TabsList, TabsTrigger } from "../ui/tabs";
 import { useState, useEffect, useRef } from "react";
 import {
   Table,
   TableBody,
   TableCell,
   TableHead,
-  TableHeader,
   TableRow,
 } from "@/components/ui/table";
 import { CheckIcon, MinusIcon } from "lucide-react";
@@ -30,8 +25,7 @@ import {
   HoverCardContent,
   HoverCardTrigger,
 } from "@/components/ui/hover-card";
-import { TrustedBy } from "./components/TrustedBy";
-import { PricingCalculator } from "./PricingCalculator";
+import { TrustedBy } from "../components/TrustedBy";
 import { trustedByData } from "@/data/trusted-by";
 
 // Reusable graduated pricing text with calculator link
@@ -48,39 +42,6 @@ const GraduatedPricingText = () => {
 };
 
 type DeploymentOption = "cloud" | "selfHosted";
-
-type DeploymentOptionParams = {
-  switch: React.ReactNode;
-  title: string;
-  subtitle: string;
-  href: string;
-};
-
-const deploymentOptions: Record<DeploymentOption, DeploymentOptionParams> = {
-  cloud: {
-    switch: (
-      <span className="flex flex-row items-center gap-x-1">
-        Langfuse Cloud
-        <span className="hidden md:block"> (we host)</span>
-      </span>
-    ),
-    title: "Pricing",
-    subtitle:
-      "Get started on the Hobby plan for free. No credit card required.",
-    href: "/pricing",
-  },
-  selfHosted: {
-    switch: (
-      <span className="flex flex-row items-center gap-x-1">
-        Self-hosted
-        <span className="hidden md:block"> (you host)</span>
-      </span>
-    ),
-    title: "Pricing",
-    subtitle: "Deploy Langfuse OSS today. Upgrade to Enterprise at any time.",
-    href: "/pricing-self-host",
-  },
-};
 
 type Tier = {
   name: string;
@@ -1255,21 +1216,266 @@ const sections: Section[] = [
   },
 ];
 
-export default function Pricing({
-  isPricingPage = false,
-  initialVariant = "cloud",
+// Helper component for feature details display
+const FeatureDetails = ({
+  description,
+  href,
 }: {
-  isPricingPage?: boolean;
-  initialVariant?: "cloud" | "selfHosted";
-}) {
-  const [localVariant, setLocalVariant] = useState(initialVariant);
-  const variant = isPricingPage ? initialVariant : localVariant;
+  description?: string;
+  href?: string;
+}) => {
+  if (!description && !href) {
+    return null;
+  }
+
+  if (!description && href) {
+    return (
+      <Link href={href} className="inline-block" target="_blank">
+        <ExternalLink className="size-4 ml-2 pt-0.5" />
+      </Link>
+    );
+  }
+
+  return (
+    <HoverCard>
+      <HoverCardTrigger>
+        <InfoIcon className="inline-block size-3 ml-1" />
+      </HoverCardTrigger>
+      <HoverCardContent className="w-60 text-xs">
+        {description}
+        {href && (
+          <span>
+            {" "}
+            (
+            <Link href={href} className="underline" target="_blank">
+              learn more
+            </Link>
+            )
+          </span>
+        )}
+      </HoverCardContent>
+    </HoverCard>
+  );
+};
+
+// Helper component for feature cell rendering
+const FeatureCell = ({
+  value,
+}: {
+  value: boolean | string | React.ReactNode;
+}) => {
+  if (typeof value === "string") {
+    return (
+      <div className="text-sm leading-6 text-center break-words">
+        {value}
+        {value.includes(TEAMS_ADDON) && (
+          <HoverCard>
+            <HoverCardTrigger>
+              <InfoIcon className="inline-block size-3 ml-1" />
+            </HoverCardTrigger>
+            <HoverCardContent className="w-60">
+              Available as part of the Teams add-on on the Pro plan.
+            </HoverCardContent>
+          </HoverCard>
+        )}
+        {value.includes(YEARLY_COMMITMENT) && (
+          <HoverCard>
+            <HoverCardTrigger>
+              <InfoIcon className="inline-block size-3 ml-1" />
+            </HoverCardTrigger>
+            <HoverCardContent className="w-60">
+              Available when committing to a yearly contract on the Enterprise
+              plan.
+            </HoverCardContent>
+          </HoverCard>
+        )}
+      </div>
+    );
+  } else if (typeof value === "boolean") {
+    return (
+      <div className="flex justify-center">
+        {value === true ? (
+          <CheckIcon className="h-5 w-5 text-primary" />
+        ) : (
+          <MinusIcon className="h-5 w-5 text-muted-foreground" />
+        )}
+      </div>
+    );
+  } else {
+    return (
+      <div className="text-sm leading-6 text-center break-words">{value}</div>
+    );
+  }
+};
+
+// Pricing Plans component - renders the card grid
+export function PricingPlans({ variant }: { variant: DeploymentOption }) {
   const selectedTiers = tiers[variant];
+
+  return (
+    <div
+      className={cn(
+        "mt-12 grid sm:grid-cols-2 gap-y-6 gap-x-6 md:gap-x-2 lg:gap-x-6 lg:items-stretch",
+        selectedTiers.length === 4 && "md:grid-cols-4",
+        selectedTiers.length === 3 && "md:grid-cols-3",
+        selectedTiers.length === 2 && "md:grid-cols-2"
+      )}
+    >
+      {selectedTiers.map((tier) => (
+        <Card
+          key={tier.id}
+          className={cn(
+            tier.featured && "border-primary",
+            "relative h-full flex flex-col"
+          )}
+        >
+          {/* Unlimited Users callout for Core and Pro */}
+          {variant === "cloud" &&
+            (tier.name === "Core" || tier.name === "Pro") && (
+              <div className="absolute -top-3 left-1/2 transform -translate-x-1/2">
+                <div className="bg-primary text-primary-foreground text-xs px-3 py-1 rounded-full font-medium text-center whitespace-nowrap inline-block">
+                  Unlimited Users
+                </div>
+              </div>
+            )}
+
+          <CardHeader className="p-4 lg:p-6 text-left">
+            <CardTitle className="text-lg text-foreground font-semibold">
+              {tier.name}
+            </CardTitle>
+            <CardDescription className="text-left">
+              {tier.description}
+              {tier.learnMore && (
+                <>
+                  {" "}
+                  <Link href={tier.learnMore} className="underline">
+                    Learn more
+                  </Link>
+                  .
+                </>
+              )}
+            </CardDescription>
+          </CardHeader>
+          <CardContent className="p-0 px-4 lg:px-6">
+            {/* Price information */}
+            <div className="h-[60px] flex items-baseline">
+              <span className="font-bold text-3xl">{tier.price}</span>
+              <span className="text-sm leading-4 ml-1">
+                {tier.price.includes("$")
+                  ? tier.priceUnit
+                    ? `/ ${tier.priceUnit}`
+                    : "/ month"
+                  : ""}
+              </span>
+            </div>
+
+            <div>
+              {tier.ctaCallout ? (
+                <div className="flex gap-2">
+                  <Button
+                    className="flex-1"
+                    variant={tier.featured ? "default" : "outline"}
+                    asChild
+                  >
+                    <Link href={tier.href}>{tier.cta}</Link>
+                  </Button>
+                  <Button className="flex-1" variant="secondary" asChild>
+                    <Link href={tier.ctaCallout.href}>
+                      {tier.ctaCallout.text}
+                    </Link>
+                  </Button>
+                </div>
+              ) : (
+                <>
+                  <Button
+                    className="w-full"
+                    variant={tier.featured ? "default" : "outline"}
+                    asChild
+                  >
+                    <Link href={tier.href}>{tier.cta}</Link>
+                  </Button>
+                </>
+              )}
+            </div>
+
+            {/* Callouts for different tiers - always render container for alignment */}
+            <div className="p-6 h-[30px] flex items-center justify-center">
+              {tier.calloutLink ? (
+                <div className="text-xs text-muted-foreground text-center">
+                  <Link
+                    href={tier.calloutLink.href}
+                    className="underline hover:text-primary"
+                  >
+                    {tier.calloutLink.text}
+                  </Link>
+                </div>
+              ) : null}
+            </div>
+          </CardContent>
+
+          {/* Trusted by section for cloud tiers */}
+          {variant === "cloud" && (
+            <>
+              <div className="border-t"></div>
+              <TrustedBy customers={trustedByData.cloud[tier.name]} />
+            </>
+          )}
+          <div className="border-t"></div>
+          <CardFooter className="p-4 lg:p-6 flex-col items-start gap-2">
+            <ul className="space-y-2.5 text-sm">
+              {tier.mainFeatures.map((feature, index) => (
+                <li key={index} className="flex space-x-2">
+                  <Check className="flex-shrink-0 mt-0.5 h-4 w-4 text-primary" />
+                  <span className="text-muted-foreground">{feature}</span>
+                </li>
+              ))}
+            </ul>
+            {tier.addOn && (
+              <div className="mt-3 border rounded pt-4 p-3 relative w-full">
+                <div className="absolute left-1/2 -translate-x-1/2 -translate-y-1/2 -top-0 bg-card px-2 text-xs text-muted-foreground">
+                  + optional
+                </div>
+                <div className="flex justify-between items-center mb-2">
+                  <span className="font-bold text-sm text-primary">
+                    {tier.addOn.name}
+                  </span>
+                  {tier.addOn.price && (
+                    <span className="font-bold text-sm text-primary">
+                      {tier.addOn.price}
+                    </span>
+                  )}
+                </div>
+                <ul className="mt-1 space-y-1 text-sm">
+                  {tier.addOn.mainFeatures.map((feature) => (
+                    <li key={feature} className="flex space-x-2">
+                      <Check className="flex-shrink-0 mt-0.5 h-4 w-4 text-primary" />
+                      <span className="text-muted-foreground">{feature}</span>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
+          </CardFooter>
+        </Card>
+      ))}
+    </div>
+  );
+}
+
+// Pricing Table component - renders the comparison tables
+export function PricingTable({
+  variant,
+  isPricingPage = false,
+}: {
+  variant: DeploymentOption;
+  isPricingPage?: boolean;
+}) {
   const [isHeaderFixed, setIsHeaderFixed] = useState(false);
   const [headerWidth, setHeaderWidth] = useState<number>(0);
   const [columnWidths, setColumnWidths] = useState<number[]>([]);
   const tableRef = useRef<HTMLDivElement>(null);
   const headerRef = useRef<HTMLTableSectionElement>(null);
+  const selectedTiers = tiers[variant];
 
   useEffect(() => {
     if (!isPricingPage) return;
@@ -1321,685 +1527,183 @@ export default function Pricing({
     };
   }, [isPricingPage]);
 
-  const FeatureDetails = ({
-    description,
-    href,
-  }: {
-    description?: string;
-    href?: string;
-  }) => {
-    if (!description && !href) {
-      return null;
-    }
-
-    if (!description && href) {
-      return (
-        <Link href={href} className="inline-block" target="_blank">
-          <ExternalLink className="size-4 ml-2 pt-0.5" />
-        </Link>
-      );
-    }
-
-    return (
-      <HoverCard>
-        <HoverCardTrigger>
-          <InfoIcon className="inline-block size-3 ml-1" />
-        </HoverCardTrigger>
-        <HoverCardContent className="w-60 text-xs">
-          {description}
-          {href && (
-            <span>
-              {" "}
-              (
-              <Link href={href} className="underline" target="_blank">
-                learn more
-              </Link>
-              )
-            </span>
-          )}
-        </HoverCardContent>
-      </HoverCard>
-    );
-  };
-
-  const FeatureCell = ({
-    value,
-  }: {
-    value: boolean | string | React.ReactNode;
-  }) => {
-    if (typeof value === "string") {
-      return (
-        <div className="text-sm leading-6 text-center break-words">
-          {value}
-          {value.includes(TEAMS_ADDON) && (
-            <HoverCard>
-              <HoverCardTrigger>
-                <InfoIcon className="inline-block size-3 ml-1" />
-              </HoverCardTrigger>
-              <HoverCardContent className="w-60">
-                Available as part of the Teams add-on on the Pro plan.
-              </HoverCardContent>
-            </HoverCard>
-          )}
-          {value.includes(YEARLY_COMMITMENT) && (
-            <HoverCard>
-              <HoverCardTrigger>
-                <InfoIcon className="inline-block size-3 ml-1" />
-              </HoverCardTrigger>
-              <HoverCardContent className="w-60">
-                Available when committing to a yearly contract on the Enterprise
-                plan.
-              </HoverCardContent>
-            </HoverCard>
-          )}
-        </div>
-      );
-    } else if (typeof value === "boolean") {
-      return (
-        <div className="flex justify-center">
-          {value === true ? (
-            <CheckIcon className="h-5 w-5 text-primary" />
-          ) : (
-            <MinusIcon className="h-5 w-5 text-muted-foreground" />
-          )}
-        </div>
-      );
-    } else {
-      return (
-        <div className="text-sm leading-6 text-center break-words">{value}</div>
-      );
-    }
-  };
-
   return (
-    <HomeSection id="pricing" className={cn(isPricingPage && "px-0 sm:px-0")}>
-      <div className="isolate overflow-hidden">
-        <div className="flow-root pb-16 lg:pb-0">
-          <div className="mx-auto max-w-7xl">
-            <Header
-              title={deploymentOptions[variant].title}
-              description={deploymentOptions[variant].subtitle}
-              h="h1"
-            />
+    <>
+      {/* Feature comparison (up to lg) */}
+      <section
+        aria-labelledby="mobile-comparison-heading"
+        className="lg:hidden mt-20"
+      >
+        <h2 id="mobile-comparison-heading" className="sr-only">
+          Feature comparison
+        </h2>
 
-            {/* Deployment Options Tabs */}
-            <Tabs
-              defaultValue={variant}
-              value={variant}
-              className="mt-4 flex justify-center"
-              onValueChange={(value) => {
-                if (!isPricingPage) {
-                  setLocalVariant(value as "cloud" | "selfHosted");
-                }
-              }}
-            >
-              <TabsList className="mx-auto">
-                {Object.keys(deploymentOptions).map((key) => (
-                  <TabsTrigger key={key} value={key} asChild={isPricingPage}>
-                    {isPricingPage ? (
-                      <Link href={deploymentOptions[key].href}>
-                        {deploymentOptions[key].switch}
-                      </Link>
-                    ) : (
-                      deploymentOptions[key].switch
-                    )}
-                  </TabsTrigger>
-                ))}
-              </TabsList>
-            </Tabs>
-
-            {/* Pricing Cards Grid */}
+        <div className="mx-auto max-w-2xl space-y-16">
+          {selectedTiers.map((tier) => (
             <div
-              className={cn(
-                "mt-12 grid sm:grid-cols-2 gap-y-6 gap-x-6 md:gap-x-2 lg:gap-x-6 lg:items-stretch",
-                selectedTiers.length === 4 && "md:grid-cols-4",
-                selectedTiers.length === 3 && "md:grid-cols-3",
-                selectedTiers.length === 2 && "md:grid-cols-2"
-              )}
+              key={tier.id}
+              className="mb-10 bg-card rounded-lg overflow-hidden border p-4"
             >
-              {selectedTiers.map((tier) => (
-                <Card
-                  key={tier.id}
-                  className={cn(
-                    tier.featured && "border-primary",
-                    "relative h-full flex flex-col"
-                  )}
-                >
-                  {/* Unlimited Users callout for Core and Pro */}
-                  {variant === "cloud" &&
-                    (tier.name === "Core" || tier.name === "Pro") && (
-                      <div className="absolute -top-3 left-1/2 transform -translate-x-1/2">
-                        <div className="bg-primary text-primary-foreground text-xs px-3 py-1 rounded-full font-medium text-center whitespace-nowrap inline-block">
-                          Unlimited Users
-                        </div>
-                      </div>
-                    )}
-
-                  <CardHeader className="p-4 lg:p-6 text-left">
-                    <CardTitle className="text-lg text-foreground font-semibold">
-                      {tier.name}
-                    </CardTitle>
-                    <CardDescription className="text-left">
-                      {tier.description}
-                      {tier.learnMore && (
-                        <>
-                          {" "}
-                          <Link href={tier.learnMore} className="underline">
-                            Learn more
-                          </Link>
-                          .
-                        </>
-                      )}
-                    </CardDescription>
-                  </CardHeader>
-                  <CardContent className="p-0 px-4 lg:px-6">
-                    {/* Price information */}
-                    <div className="h-[60px] flex items-baseline">
-                      <span className="font-bold text-3xl">{tier.price}</span>
-                      <span className="text-sm leading-4 ml-1">
-                        {tier.price.includes("$")
-                          ? tier.priceUnit
-                            ? `/ ${tier.priceUnit}`
-                            : "/ month"
-                          : ""}
-                      </span>
-                    </div>
-
-                    <div>
-                      {tier.ctaCallout ? (
-                        <div className="flex gap-2">
-                          <Button
-                            className="flex-1"
-                            variant={tier.featured ? "default" : "outline"}
-                            asChild
-                          >
-                            <Link href={tier.href}>{tier.cta}</Link>
-                          </Button>
-                          <Button
-                            className="flex-1"
-                            variant="secondary"
-                            asChild
-                          >
-                            <Link href={tier.ctaCallout.href}>
-                              {tier.ctaCallout.text}
-                            </Link>
-                          </Button>
-                        </div>
-                      ) : (
-                        <>
-                          <Button
-                            className="w-full"
-                            variant={tier.featured ? "default" : "outline"}
-                            asChild
-                          >
-                            <Link href={tier.href}>{tier.cta}</Link>
-                          </Button>
-                        </>
-                      )}
-                    </div>
-
-                    {/* Callouts for different tiers - always render container for alignment */}
-                    <div className="p-6 h-[30px] flex items-center justify-center">
-                      {tier.calloutLink ? (
-                        <div className="text-xs text-muted-foreground text-center">
-                          <Link
-                            href={tier.calloutLink.href}
-                            className="underline hover:text-primary"
-                          >
-                            {tier.calloutLink.text}
-                          </Link>
-                        </div>
-                      ) : null}
-                    </div>
-                  </CardContent>
-
-                  {/* Trusted by section for cloud tiers */}
-                  {variant === "cloud" && (
-                    <>
-                      <div className="border-t"></div>
-                      <TrustedBy customers={trustedByData.cloud[tier.name]} />
-                    </>
-                  )}
-                  <div className="border-t"></div>
-                  <CardFooter className="p-4 lg:p-6 flex-col items-start gap-2">
-                    <ul className="space-y-2.5 text-sm">
-                      {tier.mainFeatures.map((feature, index) => (
-                        <li key={index} className="flex space-x-2">
-                          <Check className="flex-shrink-0 mt-0.5 h-4 w-4 text-primary" />
-                          <span className="text-muted-foreground">
-                            {feature}
-                          </span>
-                        </li>
-                      ))}
-                    </ul>
-                    {tier.addOn && (
-                      <div className="mt-3 border rounded pt-4 p-3 relative w-full">
-                        <div className="absolute left-1/2 -translate-x-1/2 -translate-y-1/2 -top-0 bg-card px-2 text-xs text-muted-foreground">
-                          + optional
-                        </div>
-                        <div className="flex justify-between items-center mb-2">
-                          <span className="font-bold text-sm text-primary">
-                            {tier.addOn.name}
-                          </span>
-                          {tier.addOn.price && (
-                            <span className="font-bold text-sm text-primary">
-                              {tier.addOn.price}
-                            </span>
-                          )}
-                        </div>
-                        <ul className="mt-1 space-y-1 text-sm">
-                          {tier.addOn.mainFeatures.map((feature) => (
-                            <li key={feature} className="flex space-x-2">
-                              <Check className="flex-shrink-0 mt-0.5 h-4 w-4 text-primary" />
-                              <span className="text-muted-foreground">
-                                {feature}
-                              </span>
-                            </li>
-                          ))}
-                        </ul>
-                      </div>
-                    )}
-                  </CardFooter>
-                </Card>
-              ))}
+              <div className="mb-6">
+                <h4 className="text-lg text-foreground font-semibold">
+                  {tier.name}
+                </h4>
+                <p className="mt-2 text-sm text-muted-foreground">
+                  {tier.description}
+                </p>
+              </div>
+              <Table>
+                <TableBody>
+                  {sections.map((section) => (
+                    <React.Fragment key={section.name}>
+                      <TableRow className="bg-muted hover:bg-muted">
+                        <TableHead
+                          colSpan={2}
+                          className="w-full text-primary font-bold"
+                          scope="colgroup"
+                        >
+                          {section.name}
+                          <FeatureDetails
+                            description={section.description}
+                            href={section.href}
+                          />
+                        </TableHead>
+                      </TableRow>
+                      {section.features
+                        .filter((f) => variant in f.tiers)
+                        .map((feature) => (
+                          <TableRow key={feature.name}>
+                            <TableHead className="w-9/12" scope="row">
+                              {feature.name}
+                              <FeatureDetails
+                                description={feature.description}
+                                href={feature.href}
+                              />
+                            </TableHead>
+                            <TableCell className="w-3/12 text-center">
+                              <FeatureCell
+                                value={feature.tiers[variant][tier.name]}
+                              />
+                            </TableCell>
+                          </TableRow>
+                        ))}
+                    </React.Fragment>
+                  ))}
+                </TableBody>
+              </Table>
             </div>
+          ))}
+        </div>
+      </section>
 
-            {isPricingPage && (
-              <>
-                {/* Special offers callout */}
-                <div className="mt-4 mb-4 text-center">
-                  <div className="inline-flex items-center gap-6 px-4 py-2 bg-muted/30 rounded-full text-sm text-muted-foreground">
-                    <Link
-                      href="/startups"
-                      className="hover:text-primary transition-colors underline"
-                    >
-                      Startup discounts available
-                    </Link>
-                    <span className="text-muted-foreground/50">|</span>
-                    <Link
-                      href="/research"
-                      className="hover:text-primary transition-colors underline"
-                    >
-                      Academic & research benefits
-                    </Link>
-                  </div>
-                </div>
+      {/* Feature comparison (lg+) */}
+      <section
+        aria-labelledby="comparison-heading"
+        className="hidden lg:block bg-card rounded-lg overflow-hidden border mt-20"
+        ref={tableRef}
+      >
+        <h2 id="comparison-heading" className="sr-only">
+          Feature comparison
+        </h2>
 
-                {/* Feature comparison (up to lg) */}
-                <section
-                  aria-labelledby="mobile-comparison-heading"
-                  className="lg:hidden mt-20"
+        {isHeaderFixed && (
+          <div
+            className="fixed left-0 right-0 bg-muted z-50 shadow-md border-b"
+            style={{ top: "64px" }}
+          >
+            <div className="mx-auto max-w-7xl px-6">
+              <div className="overflow-hidden pl-[16px]">
+                <table
+                  className="w-full"
+                  style={{
+                    width: headerWidth,
+                  }}
                 >
-                  <h2 id="mobile-comparison-heading" className="sr-only">
-                    Feature comparison
-                  </h2>
-
-                  <div className="mx-auto max-w-2xl space-y-16">
-                    {selectedTiers.map((tier) => (
-                      <div
-                        key={tier.id}
-                        className="mb-10 bg-card rounded-lg overflow-hidden border p-4"
-                      >
-                        <div className="mb-6">
-                          <h4 className="text-lg text-foreground font-semibold">
-                            {tier.name}
-                          </h4>
-                          <p className="mt-2 text-sm text-muted-foreground">
-                            {tier.description}
-                          </p>
-                        </div>
-                        <Table>
-                          <TableBody>
-                            {sections.map((section) => (
-                              <React.Fragment key={section.name}>
-                                <TableRow className="bg-muted hover:bg-muted">
-                                  <TableHead
-                                    colSpan={2}
-                                    className="w-full text-primary font-bold"
-                                    scope="colgroup"
-                                  >
-                                    {section.name}
-                                    <FeatureDetails
-                                      description={section.description}
-                                      href={section.href}
-                                    />
-                                  </TableHead>
-                                </TableRow>
-                                {section.features
-                                  .filter((f) => variant in f.tiers)
-                                  .map((feature) => (
-                                    <TableRow key={feature.name}>
-                                      <TableHead className="w-9/12" scope="row">
-                                        {feature.name}
-                                        <FeatureDetails
-                                          description={feature.description}
-                                          href={feature.href}
-                                        />
-                                      </TableHead>
-                                      <TableCell className="w-3/12 text-center">
-                                        <FeatureCell
-                                          value={
-                                            feature.tiers[variant][tier.name]
-                                          }
-                                        />
-                                      </TableCell>
-                                    </TableRow>
-                                  ))}
-                              </React.Fragment>
-                            ))}
-                          </TableBody>
-                        </Table>
-                      </div>
-                    ))}
-                  </div>
-                </section>
-
-                {/* Feature comparison (lg+) */}
-                <section
-                  aria-labelledby="comparison-heading"
-                  className="hidden lg:block bg-card rounded-lg overflow-hidden border mt-20"
-                  ref={tableRef}
-                >
-                  <h2 id="comparison-heading" className="sr-only">
-                    Feature comparison
-                  </h2>
-
-                  {isHeaderFixed && (
-                    <div
-                      className="fixed left-0 right-0 bg-muted z-50 shadow-md border-b"
-                      style={{ top: "64px" }}
-                    >
-                      <div className="mx-auto max-w-7xl px-6">
-                        <div className="overflow-hidden pl-[16px]">
-                          <table
-                            className="w-full"
-                            style={{
-                              width: headerWidth,
-                            }}
-                          >
-                            <thead>
-                              <tr>
-                                {columnWidths.length > 0 && (
-                                  <>
-                                    <th
-                                      style={{ width: columnWidths[0] }}
-                                      className="text-left font-medium"
-                                      scope="col"
-                                    ></th>
-                                    {selectedTiers.map((tier, index) => (
-                                      <th
-                                        key={tier.id}
-                                        style={{
-                                          width: columnWidths[index + 1],
-                                        }}
-                                        className="py-2 text-center text-lg text-foreground font-semibold"
-                                        scope="col"
-                                      >
-                                        {tier.name}
-                                      </th>
-                                    ))}
-                                  </>
-                                )}
-                              </tr>
-                            </thead>
-                          </table>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  <div className="relative">
-                    <Table className="w-full">
-                      <thead ref={headerRef} className="bg-muted">
-                        <tr>
+                  <thead>
+                    <tr>
+                      {columnWidths.length > 0 && (
+                        <>
                           <th
-                            className="w-3/12 text-left font-medium"
+                            style={{ width: columnWidths[0] }}
+                            className="text-left font-medium"
                             scope="col"
                           ></th>
-                          {selectedTiers.map((tier) => (
+                          {selectedTiers.map((tier, index) => (
                             <th
                               key={tier.id}
-                              className="w-2/12 p-2 text-center text-lg text-foreground font-semibold"
+                              style={{
+                                width: columnWidths[index + 1],
+                              }}
+                              className="py-2 text-center text-lg text-foreground font-semibold"
                               scope="col"
                             >
                               {tier.name}
                             </th>
                           ))}
-                        </tr>
-                      </thead>
-                      <TableBody>
-                        {sections.map((section) => (
-                          <React.Fragment key={section.name}>
-                            <TableRow className="bg-muted/50">
-                              <TableCell colSpan={5} className="font-medium">
-                                {section.name}
-                                <FeatureDetails
-                                  description={section.description}
-                                  href={section.href}
-                                />
-                              </TableCell>
-                            </TableRow>
-                            {section.features
-                              .filter((f) => variant in f.tiers)
-                              .map((feature) => (
-                                <TableRow key={feature.name}>
-                                  <TableCell>
-                                    {feature.name}
-                                    <FeatureDetails
-                                      description={feature.description}
-                                      href={feature.href}
-                                    />
-                                  </TableCell>
-                                  {selectedTiers.map((tier) => (
-                                    <TableCell key={tier.id}>
-                                      <FeatureCell
-                                        value={
-                                          feature.tiers[variant][tier.name]
-                                        }
-                                      />
-                                    </TableCell>
-                                  ))}
-                                </TableRow>
-                              ))}
-                          </React.Fragment>
-                        ))}
-                      </TableBody>
-                    </Table>
-                  </div>
-                </section>
-              </>
-            )}
-          </div>
-        </div>
-        {isPricingPage ? (
-          <>
-            <div className="relative">
-              <div className="mx-auto max-w-7xl px-6 py-12 sm:py-16 lg:px-8 mt-16">
-                <PricingCalculator />
-                <DiscountOverview />
-                <PricingFAQ />
+                        </>
+                      )}
+                    </tr>
+                  </thead>
+                </table>
               </div>
             </div>
-          </>
-        ) : (
-          <>
-            <div className="text-center mt-10">
-              For a detailed comparison and FAQ, see our{" "}
-              <Link
-                href={deploymentOptions[variant].href}
-                className="underline"
-              >
-                pricing page
-              </Link>
-              .
-            </div>
-          </>
-        )}
-      </div>
-    </HomeSection>
-  );
-}
-
-const discounts = [
-  {
-    name: "Early-stage startups",
-    description: "50% off, first year",
-    cta: {
-      text: "Learn more and apply",
-      href: "/startups",
-    },
-  },
-  {
-    name: "Research / Students",
-    description: "Up to 100% off, limits apply",
-    cta: {
-      text: "Learn more and apply",
-      href: "/research",
-    },
-  },
-  {
-    name: "Open-source projects",
-    description: "USD 300 in credits / month, first year",
-  },
-];
-
-const DiscountOverview = () => (
-  <div className="mx-auto max-w-7xl px-6 lg:px-8 pt-8 mt-8" id="discounts">
-    <div className="mx-auto max-w-4xl">
-      <h2 className="text-2xl font-bold leading-10 tracking-tight text-primary">
-        Discounts
-      </h2>
-      <div className="mt-3 grid grid-cols-1 gap-4 sm:grid-cols-3">
-        {discounts.map((discount) => (
-          <div
-            key={discount.name}
-            className="rounded-lg border border-gray-200 p-4 shadow-sm"
-          >
-            <dt className="text-base font-semibold leading-7 text-primary">
-              {discount.name}
-            </dt>
-            <dd className="mt-2 text-sm leading-7 text-primary/60">
-              {discount.description}
-            </dd>
-            {discount.cta && (
-              <Button size="sm" variant="secondary" asChild className="mt-2">
-                <Link href={discount.cta.href} target="_blank">
-                  {discount.cta.text}
-                </Link>
-              </Button>
-            )}
           </div>
-        ))}
-      </div>
-      <p className="mt-6 leading-8 text-primary/60">
-        Reach out to{" "}
-        <Link href="mailto:support@langfuse.com" className="underline">
-          support@langfuse.com
-        </Link>{" "}
-        to apply for a discount. We want all startups, educational users,
-        non-profits and open source projects to build with Langfuse and are
-        happy to work with you to make that happen.
-      </p>
-    </div>
-  </div>
-);
+        )}
 
-const faqs = [
-  {
-    question: "What is the easiest way to try Langfuse?",
-    answer:
-      "You can view the <a class='underline' href='/demo'>public example project</a> or sign up for a <a class='underline' href='https://cloud.langfuse.com'>free account</a> to try Langfuse with your own data. The Hobby plan is completeley free and does not require a credit card.",
-  },
-  {
-    question: "Can I self-host Langfuse for free?",
-    answer:
-      "Yes, Langfuse is open source and you can self-host Langfuse for free. Use docker compose to run Langfuse locally, or use one of our templates to self-host Langfuse in production on Kubernetes. Check out the <a class='underline' href='/self-hosting'>self-hosting documentation</a> to learn more.",
-  },
-  {
-    question: "What is a billable unit?",
-    answer:
-      "A billable unit in Langfuse is any tracing data point you send to our platform - this includes the trace (a complete application interaction), observations (individual steps within a trace: Spans, Events and Generations), and scores (evaluations of your AI outputs). For a detailed explanation and an example, see our <a class='underline' href='/docs/observability/data-model'>Langfuse Data Model docs</a>.",
-  },
-  {
-    question: "How does the graduated pricing work?",
-    answer:
-      "Our graduated pricing means you pay different rates for different volume tiers. The first 100k units are included in paid plans, then you pay $8/100k for units 100k-1M, $7/100k for 1M-10M units, $6.5/100k for 10M-50M units, and $6/100k for 50M+ units. This ensures you get better rates as you scale up your usage. Use the pricing calculator to estimate your bill.",
-  },
-  {
-    question: "How can I reduce my Langfuse Cloud bill?",
-    answer:
-      "The primary way to reduce your Langfuse Cloud bill is to reduce the number of billable units that you ingest. We have summarized how this can be done <a class='underline' href='/faq/all/cutting-costs'>here</a>. Additionally, with our new graduated pricing model, you automatically get lower rates per 100k units as your volume increases.",
-  },
-  {
-    question: "When do I get billed?",
-    answer:
-      "You get one bill each month. We charge your Core, Pro, or Team plan at the start of the month. We charge for your usage at the end of the month. The bill you get at the start of the month shows two things: the plan cost for the new month and the usage from last month.",
-  },
-  {
-    question: "How can I manage my subscription?",
-    answer:
-      "You can manage your subscription through the organization settings in Langfuse Cloud or by using this <a class='underline' href='/billing-portal'>Customer Portal</a>.",
-  },
-  {
-    question: "Can I redline the contracts?",
-    answer:
-      "Yes, we offer customized contracts for Langfuse Enterprise customers with a yearly commitment. Please contact us at enterprise@langfuse.com for more details. The default plans are affordable as they are designed to be self-serve on our standard terms.",
-  },
-  {
-    question: "Where is the data stored?",
-    answer:
-      "Langfuse Cloud is hosted on AWS and data is stored in the US or EU depending on your selection. See our <a class='underline' href='/security'>security and privacy documentation</a> for more details.",
-  },
-  {
-    question: "I have security questions, where to start?",
-    answer:
-      "We publish almost all of our security documentation and controls. Please refer to our <a class='underline' href='/security'>security documentation</a> for more details.",
-  },
-];
-
-export function PricingFAQ() {
-  return (
-    <div id="faq">
-      <div className="mx-auto max-w-7xl px-6 pb-24 pt-16 lg:px-8">
-        <div className="mx-auto max-w-4xl">
-          <h2
-            className="text-2xl font-bold leading-10 tracking-tight 
-          text-primary"
-          >
-            Frequently asked questions
-          </h2>
-          <dl className="mt-3 space-y-0 divide-y divide-primary/10">
-            {faqs.map((faq) => (
-              <Disclosure as="div" key={faq.question}>
-                {({ open }) => (
-                  <>
-                    <dt>
-                      <Disclosure.Button className="flex w-full items-start justify-between text-left text-primary py-2.5">
-                        <span className="text-[15px] font-medium leading-6">
-                          {faq.question}
-                        </span>
-                        <span className="ml-6 flex h-6 items-center">
-                          {open ? (
-                            <Minus className="h-4 w-4" aria-hidden="true" />
-                          ) : (
-                            <Plus className="h-4 w-4" aria-hidden="true" />
-                          )}
-                        </span>
-                      </Disclosure.Button>
-                    </dt>
-                    <Disclosure.Panel as="dd" className="pb-2.5 pr-12">
-                      <p
-                        className="text-sm leading-7 text-primary/60"
-                        dangerouslySetInnerHTML={{ __html: faq.answer }}
+        <div className="relative">
+          <Table className="w-full">
+            <thead ref={headerRef} className="bg-muted">
+              <tr>
+                <th className="w-3/12 text-left font-medium" scope="col"></th>
+                {selectedTiers.map((tier) => (
+                  <th
+                    key={tier.id}
+                    className="w-2/12 p-2 text-center text-lg text-foreground font-semibold"
+                    scope="col"
+                  >
+                    {tier.name}
+                  </th>
+                ))}
+              </tr>
+            </thead>
+            <TableBody>
+              {sections.map((section) => (
+                <React.Fragment key={section.name}>
+                  <TableRow className="bg-muted/50">
+                    <TableCell colSpan={5} className="font-medium">
+                      {section.name}
+                      <FeatureDetails
+                        description={section.description}
+                        href={section.href}
                       />
-                    </Disclosure.Panel>
-                  </>
-                )}
-              </Disclosure>
-            ))}
-          </dl>
+                    </TableCell>
+                  </TableRow>
+                  {section.features
+                    .filter((f) => variant in f.tiers)
+                    .map((feature) => (
+                      <TableRow key={feature.name}>
+                        <TableCell>
+                          {feature.name}
+                          <FeatureDetails
+                            description={feature.description}
+                            href={feature.href}
+                          />
+                        </TableCell>
+                        {selectedTiers.map((tier) => (
+                          <TableCell key={tier.id}>
+                            <FeatureCell
+                              value={feature.tiers[variant][tier.name]}
+                            />
+                          </TableCell>
+                        ))}
+                      </TableRow>
+                    ))}
+                </React.Fragment>
+              ))}
+            </TableBody>
+          </Table>
         </div>
-      </div>
-    </div>
+      </section>
+    </>
   );
 }
