+++ pages/pricing.mdx
@@ -2,6 +2,6 @@
 description: Simple pricing for projects of all sizes. Get started on our Hobby plan without a credit card.
 ---
 
-import Pricing from "@/components/home/Pricing";
+import { PricingPage } from "@/components/home/pricing";
 
-<Pricing isPricingPage />
+<PricingPage isPricingPage />
